This is a P5 project to evolve a genetic code an draw the result.
The algorithm starts with a few starting genes and has a map telling how
each genes evolves into other genes. By applying these evolutions over a
few generation, we have a final sequence of genes. Then each genes is
mapped to a line direction and we draw the final DNA.

The genes are letters (a,b, c, ...) and the evolution are given by
specifying how a letter evolves, by separating the base gene to its
targets gene sequence with a column, like this for example: a:cda b:ba.
The drawing direction is given as angles in a similar format, like this
for example: a:90 b:45 c:-90.

Created a create-ui system that analyze a JS object and generates the UI
to edit its variables with support for strings, numbers, true/false
flags, choices (arrays), action buttons. Support calling a callback when
the user edits the value. The choice UI allows creating new entries or
saving a modified entry.

Create a canvas-movement system to translate and zoom the canvas with
the mouse. Translation is done by dragging, zooming withthe mouse wheel.

The evolution and drawing is done with a multi-steps algorithm so that
the user interface stays responsive.

The evolution algorithm is optimized by calculating all generations at
once by going depth-first into the generation and genetic evolution.
This allows single-pass, incremental drawing of the result. The drawing
of genes is queued as batches f one thousand segments.